requirements.txt
# requirements.txt
# Core dependencies
opencv-python>=4.8.0
ultralytics>=8.0.0
python-telegram-bot>=20.0
numpy>=1.21.0
sqlite3

# GPS and location
pyserial>=3.5
geopy>=2.3.0

# Configuration and environment
python-dotenv>=1.0.0

# Logging
logging

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0